lib/Moo/HandleMoose.pm: Possible work around for rt#84615
If the constructor autovivification call returns 'undef', then the
package we're trying to get the contructor for, ( and from that, the
    constructor specs ), that package was not constructed via

  package;
  use Moo;

doesn't turn up in %Moo::MAKERS, and must either be either Moose::Object itself, or a simple child class of
that.

Therefore, assume it has no attributes, and inflate a minimal
methods-only metaclass.